pyproject.toml
[project]
name = "cloud-cost-intelligence"
version = "0.1.0"
description = "Cloud cost intelligence platform with live AWS/GCP ingestion, anomaly detection, and automated optimizations."
requires-python = ">=3.11"
dependencies = [
  "apscheduler>=3.10.4",
  "boto3>=1.34.0",
  "email-validator>=2.2.0",
  "fastapi>=0.110.0",
  "google-api-python-client>=2.137.0",
  "google-cloud-bigquery>=3.25.0",
  "google-cloud-monitoring>=2.22.0",
  "httpx>=0.27.0",
  "itsdangerous>=2.2.0",
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "pydantic-settings>=2.3.4",
  "python-dateutil>=2.9.0.post0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "scikit-learn>=1.5.1",
  "sqlalchemy>=2.0.31",
  "uvicorn>=0.30.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2",
  "pytest-asyncio>=0.23.7",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
